+++ sshd-core/src/main/java/org/apache/sshd/common/channel/AbstractChannel.java
@@ -950,6 +950,11 @@ public abstract class AbstractChannel
     }
 
     @Override
+    public int getAttributesCount() {
+        return attributes.size();
+    }
+
+    @Override
     @SuppressWarnings("unchecked")
     public <T> T getAttribute(AttributeRepository.AttributeKey<T> key) {
         return (T) attributes.get(Objects.requireNonNull(key, "No key"));
@@ -979,6 +984,11 @@ public abstract class AbstractChannel
     @SuppressWarnings("unchecked")
     public <T> T removeAttribute(AttributeRepository.AttributeKey<T> key) {
         return (T) attributes.remove(Objects.requireNonNull(key, "No key"));
+    }
+
+    @Override
+    public void clearAttributes() {
+        attributes.clear();
     }
 
     protected void configureWindow() {