+++ sshd-core/src/main/java/org/apache/sshd/common/helpers/AbstractFactoryManager.java
@@ -148,6 +148,11 @@ public abstract class AbstractFactoryManager extends AbstractKexFactoryManager i
     }
 
     @Override
+    public int getAttributesCount() {
+        return attributes.size();
+    }
+
+    @Override
     @SuppressWarnings("unchecked")
     public <T> T getAttribute(AttributeRepository.AttributeKey<T> key) {
         return (T) attributes.get(Objects.requireNonNull(key, "No key"));
@@ -177,6 +182,11 @@ public abstract class AbstractFactoryManager extends AbstractKexFactoryManager i
     @SuppressWarnings("unchecked")
     public <T> T removeAttribute(AttributeRepository.AttributeKey<T> key) {
         return (T) attributes.remove(Objects.requireNonNull(key, "No key"));
+    }
+
+    @Override
+    public void clearAttributes() {
+        attributes.clear();
     }
 
     @Override